+++ packages/node-core/src/indexer/store.service.ts
@@ -403,7 +403,10 @@ group by
     return (
       this.modelIndexedFields.findIndex(
         (indexField) =>
-          upperFirst(camelCase(indexField.entityName)) === entity && camelCase(indexField.fieldName) === field
+          (upperFirst(camelCase(indexField.entityName)) === entity || indexField.entityName === entity) &&
+          // We add this because in some case upperFirst and camelCase will not match with entity name,
+          // see test entity name like `MinerIP`
+          camelCase(indexField.fieldName) === field
       ) > -1
     );
   }